[SW-69] fix: ListBarSlice로 네이밍 수정
