+++ frontend/src/store/store.js
@@ -1,13 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit';
 import authSlice from './slices/authSlice';
-import sideBarSlice from './slices/sideBarSlice';
+import listBarSlice from './slices/listBarSlice';
 import userSlice from './slices/userSlice';
-import detailViewSlice from './slices/detailViewSlice'
-import kakaoMapSlice from './slices/kakaoMapSlice'
+import detailViewSlice from './slices/detailViewSlice';
+import kakaoMapSlice from './slices/kakaoMapSlice';
+
 const store = configureStore({
   reducer: {
     auth: authSlice,
-    sideBar: sideBarSlice,
+    listBar: listBarSlice,
     detailView: detailViewSlice,
     kakaoMap: kakaoMapSlice,
     user: userSlice,
